Extract rpc logic into a separate module
